# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OS Manager API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor Local"
        }
    ],
    "paths": {
        "/api/configuracoes": {
            "get": {
                "tags": [
                    "Configuracoes"
                ],
                "summary": "Retorna as configurações do sistema",
                "operationId": "ced427248dd5762dc267b139c6aa4e69",
                "responses": {
                    "200": {
                        "description": "Configurações retornadas"
                    },
                    "403": {
                        "description": "Acesso negado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Configuracoes"
                ],
                "summary": "Atualiza as configurações do sistema",
                "operationId": "f453d0388275c9de6006d0cecf8d1d0d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "sla_muito_alta": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "sla_alta": {
                                        "type": "integer",
                                        "example": 4
                                    },
                                    "sla_media": {
                                        "type": "integer",
                                        "example": 8
                                    },
                                    "sla_baixa": {
                                        "type": "integer",
                                        "example": 24
                                    },
                                    "nome_sistema": {
                                        "type": "string",
                                        "example": "Central de Suporte"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Configurações salvas"
                    },
                    "422": {
                        "description": "Erro de validação"
                    },
                    "403": {
                        "description": "Acesso negado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/notificacoes": {
            "get": {
                "tags": [
                    "Notificacoes"
                ],
                "summary": "Lista Notificações",
                "description": "Retorna todas as notificações do usuário logado, ordenadas das mais recentes para as mais antigas.",
                "operationId": "a556d23d6c230e3bb85a27cc91085a78",
                "responses": {
                    "200": {
                        "description": "Lista de notificações retornada com sucesso"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/notificacoes/{id}/ler": {
            "put": {
                "tags": [
                    "Notificacoes"
                ],
                "summary": "Marcar Notificação como Lida",
                "description": "Marca uma notificação específica do usuário logado como lida.",
                "operationId": "409edd847fc521f565db8b502c0d3cc6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da notificação",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notificação marcada como lida"
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "404": {
                        "description": "Notificação não encontrada ou não pertence ao usuário"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/notificacoes/ler-todas": {
            "put": {
                "tags": [
                    "Notificacoes"
                ],
                "summary": "Marcar Todas as Notificações como Lidas",
                "description": "Marca todas as notificações não lidas do usuário logado como lidas.",
                "operationId": "0ea7010324f116e669e8ed94c6ba46be",
                "responses": {
                    "200": {
                        "description": "Todas as notificações marcadas como lidas"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}